Index label related transformations

.t() allows to substitute index labels, raising and lowering indices and perform Kronecker products in any number or order. Multiple transformations are separated by a comma ",".

Usage

.t(x, ..., g = getOption("ricci.g"))

Arguments

x

A labeled tensor object, created by %_% or tensor().

...

Any number of transformation specifications separated by a comma.

g

An (unlabeled) covariant metric tensor as array used for raising and lowering indices.

Value

A modified tensor object.

Details

Substituting index label (without altering position) <label1> -> <label2>

Raising index: -<label1> -> +<label1>

Lowering index: +<label1> -> -<label2>

Kronecker product: .(<label1>, <label2>, ..., <labeln-1>) -> <labeln+1>

Examples

a <- array(1:8, dim = c(2, 2, 2))

# raising an index and contracting
a %_% .(i, j, k) |>
  .t(-j -> +i)
#> <Labeled Tensor> [2x2x2] / .(-j, -k, +i)

# taking the Kronecker product
a %_% .(i, j, k) |>
  .t(.(j, k) -> l)
#> <Labeled Tensor> [4x2] / .(-l, -i)
#>      [,1] [,2]
#> [1,]    1    2
#> [2,]    3    4
#> [3,]    5    6
#> [4,]    7    8